feat: Convert to multi-page application with routing
This commit refactors the single-page application into a multi-page application using `react-router-dom`.

The main changes include:
- Addition of `react-router-dom@6` to the project.
- Creation of a `pages` directory with components for Home, About, Courses, and Contact.
- Creation of a `components` directory with `Header`, `Footer`, and `Layout` components.
- The main `App.js` has been updated to handle the routing logic.
- The navigation has been updated to use `Link` components for client-side routing.

The 3D design part of the task has been postponed to a future update, as it requires a major version upgrade of React.

Note: I could not run the test suite due to a persistent npm environment issue.
