+++ package.json
@@ -10,6 +10,7 @@
     "axios": "^0.21.1",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
+    "react-router-dom": "^6.30.1",
     "react-scripts": "4.0.3",
     "web-vitals": "^1.1.2"
   },
