+++ src/pages/About.js
@@ -1,0 +1,29 @@
+import React from 'react';
+import '../App.css';
+
+const About = () => {
+    return (
+        <section id="about-section">
+            <div id="about-section1">
+                <h3 style={{ color: "#333", fontSize: "35px" }}>About Us</h3>
+                <p>
+                    Welcome to Perseverance Software Training Institute, located in the vibrant city of Tirupati. Our institute is committed to providing top-notch education in the field of software development and IT.
+                </p>
+                <p>
+                    At Perseverance, we believe in fostering a learning environment that encourages creativity, critical thinking, and hands-on experience. Our dedicated team of instructors brings industry expertise to the classroom, ensuring that our students are well-prepared for the challenges of the IT field.
+                </p>
+                <p>
+                    Whether you're looking to start a career in DevOps, Cloud Computing, Programming, or other IT domains, Perseverance is here to guide you on your journey. Join us and become a part of a community that values perseverance, dedication, and continuous learning.
+                </p>
+                <p>
+                    Join us on a journey of knowledge and skill development. Whether you are a beginner looking to start your career in IT or a professional seeking to enhance your skills, Perseverance is the place for you.
+                </p>
+                <p>
+                    Come, be a part of our community, and let's build a successful future together.
+                </p>
+            </div>
+        </section>
+    );
+};
+
+export default About;
