+++ src/pages/Contact.js
@@ -1,0 +1,16 @@
+import React from 'react';
+import '../App.css';
+
+const Contact = () => {
+    return (
+        <section id="contact-section">
+            <h3 style={{ color: "#333", fontSize: "35px" }}>Contact Us</h3>
+            <div className="contact-info">
+                <p>Phone: 7075505229, 7075506229</p>
+                <p>Email: <a href="mailto:persevcareers@gmail.com">persevcareers@gmail.com</a></p>
+            </div>
+        </section>
+    );
+};
+
+export default Contact;
